Add a benchmark for round-tripping an error code through EH.
In re a discussion on the SG14 mailing list regarding the
potential cost of "magically" converting errors to exceptions,
resp. vice versa, at API boundaries which cannot propagate errors,
resp. exceptions.

On my MacBook, the results are:

    -rw-r--r--  1 ajo  staff  1944 Apr 11 14:58 eh.o
    -rw-r--r--  1 ajo  staff   988 Apr 11 14:58 noeh.o

That's 1944 bytes of .o for round-tripping, versus 988 bytes
of .o for doing the whole thing natively in error codes.